Expose line no. information for each statement
This patch adds support for storing the start/end line number for each
statement present in the original input. The CLI only prints this
information when the JSON formatter is used, as of now:

    [
      {
        "sql": "......",
        "locks_acquired": [],
        "location": {
          "start_line": 1,
          "end_line": 3
        }
    ]

To do this, we migrate from libpq_query to sqlparser, which expose line
no. information. There's still a small bug reported upstream[1] but this
is at least a good starting point.

Part of #17

[1] https://github.com/apache/datafusion-sqlparser-rs/issues/1858